+++ src/main/java/com/nhomX/example/utils/CurrencyFormatter.java
@@ -4,13 +4,13 @@ import java.text.DecimalFormat;
 
 public class CurrencyFormatter {
     // Định dạng xèng
-    public static String formatVND(double amount){
+    public static String formatVND(long amount){
         DecimalFormat formatter = new DecimalFormat("#,###");
         String formatterNumber = formatter.format(amount).replace(",",".");
         return formatterNumber + " VNĐ";
     }
     // Định dạng so:
-    public static String formatNumber(double amount){
+    public static String formatNumber(long amount){
         DecimalFormat formatter = new DecimalFormat("#,###");
         return formatter.format(amount).replace(",",".");
     }
